Dialog searching basically working
Got the dialog searching up and working. Although it only matches exact word for word quotes, it works surprisingly well. When a match is found it generates a VLC link that allows you to launch the movie at the point where that quote is said. The first few times it worked I just laughed, its so damn cool.

Next up will be to integrate these subtitles results into the search results page as well as implement more fuzzy searching.

+++ src/logic/subtitles.py
@@ -1,10 +1,11 @@
 #All the subtitle related code here
 import json
+import re
 from subprocess import Popen, PIPE
 import subprocess
 from base64 import urlsafe_b64encode
 from PyQt5.QtCore import QThread, QObject, pyqtSignal, QVariant
-from .movie_library import checkDbOpen, getDbCursor
+from .movie_library import checkDbOpen, getDbCursor, fixDbData
 from dialogs.widgets.generic_progress_bar import genericProgressDialog
 
 import os, sqlite3
@@ -12,12 +13,23 @@ import os, sqlite3
 SUBLANG = "eng"
 DATABASE_PATH = "./subtitles.db"
 
+
 class SubtitleLibrary:
     def __init__(self, DBPATH=DATABASE_PATH):
         self.dbmutex = False
         self.subsdict = {}
         self.dbpath = DBPATH
         self.checkForSubtitleDb()
+        self.fieldlist = self.getFieldList()
+
+    @checkDbOpen
+    def getFieldList(self):
+        with getDbCursor(self.dbpath, self.dbmutex) as dbcursor:
+            data = dbcursor.execute("PRAGMA table_info(subtitle_data)").fetchall()
+        fieldlist = []
+        for c in data:
+            fieldlist.append(c[1])
+        return fieldlist
 
 
     @checkDbOpen
@@ -49,7 +61,8 @@ class SubtitleLibrary:
 
     @checkDbOpen
     def checkForSubs(self, movietitle):
-        if len(movietitle) == 0 or not isinstance(movietitle, str): return False
+        if len(movietitle) == 0 or not isinstance(movietitle, str):
+            return False
         with getDbCursor(self.dbpath, self.dbmutex) as dbcursor:
             data = dbcursor.execute("SELECT COUNT(1) FROM subtitle_data WHERE title=?", (movietitle,)).fetchone()
         return bool(data[0])
@@ -74,6 +87,55 @@ class SubtitleLibrary:
                                  )
             self.subsdict[subsdata["title"]] = subsdata
 
+    #Simple search, just returning subdata for whatever movies are in movielist
+    def search(self, movielist):
+        if len(movielist) == 0: return
+        #Subtitles are stored in their raw srt format so we can't use the SQL query to search for the
+        #dialogstr with any actual reliability. Have to get all the subs first, then search.
+        querystr = "SELECT * FROM subtitle_data WHERE title=\"%s\"" % movielist[0]
+        for m in movielist[1:]:
+            querystr += " OR title=\"%s\"" % m
+        returndata = self._SEARCH(querystr)
+        if len(returndata) == 0:
+            print("No subtitles in the database for: \n%s" % ", ".join(movielist))
+            return {}
+        #Now we can convert the raw srt format into a single corpus we can search more easily.
+        #We still want to be able to reference the timecode our dialog match came from so
+        #as we build our corpus we will keep track of what index each timecode's dialog is
+        #added at. Later we can reference the index of the dialog to find the timecode again.
+        for subdata in returndata.values():
+            #All the srt files I have seen so far delineate separate pieces of dialog by an empty line.
+            splitsubs = re.split(r"\n^$\n", subdata["subtitles"], flags=re.MULTILINE)
+            subcorpus = ""
+            timecodes = {} # key=index , val = timecode
+            for line in splitsubs:
+                if len(line.strip()) == 0:
+                    continue
+                srtregex = r"([0-9]{2}:[0-9]{2}:[0-9]{2},[0-9]{3}\s*-->\s*[0-9]{2}:[0-9]{2}:[0-9]{2},[0-9]{3})(.*)"
+                tcmatch = re.search(srtregex, line, flags=re.MULTILINE|re.DOTALL)
+                if tcmatch is None:
+                    print("Found an invalid line in file %s. Is this a valid srt format?" % subdata["filename"])
+                    continue
+                timecode = tcmatch.group(1)
+                dialog = tcmatch.group(2).strip()
+                #Some srt files include html tags so we'll remove those as well.
+                #I havent seen any complicated tags yet so for now a regex is fine.
+                dialog = re.sub(r"<[\s\w/]+?>", "", dialog, flags=re.M)
+                #Also remove any newlines
+                dialog = re.sub(r"\n", " ", dialog, flags=re.M)
+                index = len(subcorpus)
+                subcorpus += dialog + " "
+                timecodes[index] = timecode
+            subdata["subtitles"] = {"timecodes": timecodes, "corpus": subcorpus}
+        return returndata
+
+    @checkDbOpen
+    def _SEARCH(self, querystr):
+        with getDbCursor(self.dbpath, self.dbmutex) as dbcursor:
+            return fixDbData(dbcursor.execute(querystr).fetchall(), self.fieldlist)
+
+
+
 class SubtitleExtractor(QObject):
     newSubtitles = pyqtSignal(QVariant)
     threadFinished = pyqtSignal(QVariant)
@@ -82,6 +144,7 @@ class SubtitleExtractor(QObject):
         super(SubtitleExtractor, self).__init__(parent)
         self.moviedata = moviedata
         self.cmd = None
+        self.stopping = False
 
     def getSubs(self):
         if "substmp" not in os.listdir():
@@ -89,19 +152,22 @@ class SubtitleExtractor(QObject):
         #Get track id, extract track to file, read file data into dict, emit data
         vidpath = self.moviedata["filelocation"].replace(r"\\", r"\\") + "\\" + self.moviedata["filename"].replace(r"\\", r"\\")
         trackid = self.getSubtitleTrackIds(vidpath)
-        if len(trackid) == 0: #Didnt find subs
+        if self.stopping: return
+        if trackid is None or len(trackid) == 0: #Didnt find subs
             self.threadFinished.emit(self)
             print("NOFINDSUBS")
             return
         trackid = trackid[0]
         rawsubs = self.extractSubtitlesFromVideo(vidpath, trackid, self.moviedata["title"])
+        if self.stopping: return
         #Emitdata is just the data we need for addSubs()
-        emitdata = {}
-        emitdata["title"] = self.moviedata["title"]
-        emitdata["filename"] = self.moviedata["filename"]
-        emitdata["filelocation"] = self.moviedata["filelocation"]
-        emitdata["subtitles"] = rawsubs
-        emitdata["extra1"] = ""
+        emitdata = {
+            "title": self.moviedata["title"],
+            "filename": self.moviedata["filename"],
+            "filelocation": self.moviedata["filelocation"],
+            "subtitles": rawsubs,
+            "extra1": ""
+        }
         self.newSubtitles.emit(emitdata)
         self.threadFinished.emit(self)
 
@@ -110,6 +176,7 @@ class SubtitleExtractor(QObject):
         cmd = "mkvmerge --identification-format json --identify \"%s\"" % vidpath
         self.cmd = Popen(cmd, stdout=PIPE)
         rawout, errors = self.cmd.communicate()
+        if self.stopping: return
         videodata = json.loads(rawout)
         goodsubids = []
         if "tracks" not in videodata: #Something bigly wrong
@@ -136,6 +203,7 @@ class SubtitleExtractor(QObject):
         cmd = "mkvextract tracks \"%s\" %s:%s" % (vidpath, trackid, subsfilename)
         self.cmd = Popen(cmd, stdout=PIPE)
         out, err = self.cmd.communicate()
+        if self.stopping: return
         with open(subsfilename, "r", encoding="utf8") as subsfile:
             subtitles = subsfile.read()
         os.remove(subsfilename)
@@ -153,6 +221,7 @@ class SubtitleDownloader(QObject):
         self.moviequeue = []
         self.addedsubs = 0
         self.stopping = False
+        self.progressbar = None
 
     def goodSubsCallback(self, subsdata):
         #Ignore any sub callbacks if we are stopping as its probably incomplete
@@ -188,8 +257,10 @@ class SubtitleDownloader(QObject):
         self.progressbar.closableDialogClosing.connect(self.progressbar.accept)
         self.stopping = True
         for t in self.threadlist.values():
-            c = subprocess.Popen("taskkill /F /T /PID %s" % t.cmd.pid, stdout=PIPE)
+            t.stopping = True
+            _ = subprocess.Popen("taskkill /F /T /PID %s" % t.cmd.pid, stdout=PIPE)
             t.threadref.quit()
+        self.progressbar.setFinished()
 
     def updateSubsCache(self):
         #Spin up our progress bar, go through library and spin off threads for extract operations
@@ -208,6 +279,8 @@ class SubtitleDownloader(QObject):
             self.createExtractionThread()
 
     def createExtractionThread(self):
+        if len(self.moviequeue) == 0:
+            return
         title = self.moviequeue.pop(0)
         moviedata = self.movieslib[title]
         while self.subdbref.checkForSubs(title) is True: #Keep popping new titles until we get one we need to create a thread for
